Get columns count on millesimes

This command return the list of columns founded in the input tibble with the count. It allows you to verify millesimes are coherent

Usage

columns(data, quiet = FALSE)

Arguments

data

a datafiles tibble (as returned by the datafiles() function)

quiet

logical; suppress message. Default to FALSE

Value

a tibble with columns found in input datafiles and occurrence for each column

Examples

datafiles() %>% columns()
#> nb of datafiles: 124
#> # A tibble: 723 × 6
#>    name  description type   format unit  `nb of occurrences`
#>    <chr> <chr>       <chr>  <chr>  <chr>               <int>
#>  1 A1955 A1955       number NA     n/a                     4
#>  2 A1960 A1960       number NA     n/a                     4
#>  3 A1965 A1965       number NA     n/a                     4
#>  4 A1970 A1970       number NA     n/a                     4
#>  5 A1975 A1975       number NA     n/a                     4
#>  6 A1980 A1980       number NA     n/a                     4
#>  7 A1982 A1982       number NA     n/a                    25
#>  8 A1983 A1983       number NA     n/a                    25
#>  9 A1984 A1984       number NA     n/a                    25
#> 10 A1985 A1985       number NA     n/a                    25
#> # … with 713 more rows