Get data

Get the data of the last millesime of all the datafiles found in data. All columns are returned as chr (see below).

get_data(data, query = list(), col_types = cols(.default = "c"), concat = TRUE)

Arguments

data

a tibble issued from datafiles() or a dataframe with two columns ridandmillesime`.

query

a query to pass to the API to select columns and filter on values.

col_types

how to convert columns, the default is to use char for all columns cols(.default = "c")

concat

TRUE

Value

If concat is TRUE (default), return a tibble with all data concatenated in one tibble. If concat is FALSE, return a list of tibbles.

get_data() returns only chr

Details

You can use convert() to convert number and integer.

For private life reason, data returned by DiDo can be secretize (the value is replaced by the string "secret") so readr can't determine data type.

Examples

# get all columns
datafiles() %>%
  dido_search("drom") %>%
  get_data()
#> # A tibble: 55 × 10
#>    DEPART…¹ DEPAR…² REGIO…³ REGIO…⁴ ANNEE ESSEN…⁵ GAZOL…⁶ FIOUL…⁷ GPL_M3 CARBU…⁸
#>    <chr>    <chr>   <chr>   <chr>   <chr> <chr>   <chr>   <chr>   <chr>  <chr>  
#>  1 971      Guadel… 01      Guadel… 2021  118664  190809  246583  20837  107177 
#>  2 972      Martin… 02      Martin… 2021  125134  171964  377837  17536  88983  
#>  3 973      Guyane  03      Guyane  2021  39528   106097  78464   8826   35147  
#>  4 974      La Réu… 04      La Réu… 2021  151834  494126  266309  38040  192811 
#>  5 976      Mayotte 06      Mayotte 2021  22575   133408  0       11427  7061   
#>  6 971      Guadel… 01      Guadel… 2020  112378  187242  273707  21455  96587  
#>  7 972      Martin… 02      Martin… 2020  115415  167795  372927  17777  77281  
#>  8 973      Guyane  03      Guyane  2020  26717   82004   106901  9147   28781  
#>  9 974      La Réu… 04      La Réu… 2020  129677  445872  202724  38551  168824 
#> 10 976      Mayotte 06      Mayotte 2020  19592   124040  0       10852  6044   
#> # … with 45 more rows, and abbreviated variable names ¹​DEPARTEMENT_CODE,
#> #   ²​DEPARTEMENT_LIBELLE, ³​REGION_CODE, ⁴​REGION_LIBELLE, ⁵​ESSENCE_M3,
#> #   ⁶​GAZOLE_M3, ⁷​FIOUL_M3, ⁸​CARBUREACTEUR_M3
# get only DEPARTEMENT_CODE and ESSENCE_M3 columns
datafiles() %>%
  dido_search("drom") %>%
  get_data(query = c(columns = "DEPARTEMENT_CODE,ESSENCE_M3"))
#> # A tibble: 55 × 2
#>    DEPARTEMENT_CODE ESSENCE_M3
#>    <chr>            <chr>     
#>  1 971              118664    
#>  2 972              125134    
#>  3 973              39528     
#>  4 974              151834    
#>  5 976              22575     
#>  6 971              112378    
#>  7 972              115415    
#>  8 973              26717     
#>  9 974              129677    
#> 10 976              19592     
#> # … with 45 more rows
# get only rows where `DEPARTEMENT_CODE == 971
datafiles() %>%
  dido_search("drom") %>%
  get_data(query = c(DEPARTEMENT_CODE = "eq:971"))
#> # A tibble: 11 × 10
#>    DEPART…¹ DEPAR…² REGIO…³ REGIO…⁴ ANNEE ESSEN…⁵ GAZOL…⁶ FIOUL…⁷ GPL_M3 CARBU…⁸
#>    <chr>    <chr>   <chr>   <chr>   <chr> <chr>   <chr>   <chr>   <chr>  <chr>  
#>  1 971      Guadel… 01      Guadel… 2021  118664  190809  246583  20837  107177 
#>  2 971      Guadel… 01      Guadel… 2020  112378  187242  273707  21455  96587  
#>  3 971      Guadel… 01      Guadel… 2019  128464  217934  233765  21539  160846 
#>  4 971      Guadel… 01      Guadel… 2018  121946  215904  215140  19534  156785 
#>  5 971      Guadel… 01      Guadel… 2017  121076  216676  204647  22433  148860 
#>  6 971      Guadel… 01      Guadel… 2016  131545  217084  217429  19206  135251 
#>  7 971      Guadel… 01      Guadel… 2015  127784  225313  190064  22612  130018 
#>  8 971      Guadel… 01      Guadel… 2014  112856  202485  37966   21274  125831 
#>  9 971      Guadel… 01      Guadel… 2013  117255  203971  209534  21946  127759 
#> 10 971      Guadel… 01      Guadel… 2012  125695  221092  324324  21934  127180 
#> 11 971      Guadel… 01      Guadel… 2011  123811  240857  336190  21589  132008 
#> # … with abbreviated variable names ¹​DEPARTEMENT_CODE, ²​DEPARTEMENT_LIBELLE,
#> #   ³​REGION_CODE, ⁴​REGION_LIBELLE, ⁵​ESSENCE_M3, ⁶​GAZOLE_M3, ⁷​FIOUL_M3,
#> #   ⁸​CARBUREACTEUR_M3