convert

Try to convert columns according to types sended by the API.

convert(data)

Arguments

data

a dataframe returned by get_data()

Value

a tibble with converted integer and number columns

Details

CAUTION For private life reason, data returns by DiDo are secretize (the value is replaced by the string "secret") so readr can't determine data type.

All secret values will be transform to NA

You can find column's type with columns()

Examples

datafiles() %>%
  dido_search("6c79805c-def9-4695-9d9f-7d86599c4d8a") %>%
  get_data() %>%
  convert()
#> # A tibble: 50 × 10
#>    DEPARTEMENT_CODE DEPARTEMENT_LIB… REGION_CODE REGION_LIBELLE ANNEE ESSENCE_M3
#>    <chr>            <chr>            <chr>       <chr>          <chr>      <dbl>
#>  1 972              Martinique       02          Martinique     2015      110986
#>  2 974              La Réunion       04          La Réunion     2019      142206
#>  3 976              Mayotte          06          Mayotte        2019       20514
#>  4 971              Guadeloupe       01          Guadeloupe     2017      121076
#>  5 973              Guyane           03          Guyane         2017       34182
#>  6 976              Mayotte          06          Mayotte        2016       20025
#>  7 973              Guyane           03          Guyane         2011       25111
#>  8 973              Guyane           03          Guyane         2014       32161
#>  9 974              La Réunion       04          La Réunion     2014      122604
#> 10 976              Mayotte          06          Mayotte        2011       16176
#> # … with 40 more rows, and 4 more variables: GAZOLE_M3 <dbl>, FIOUL_M3 <dbl>,
#> #   GPL_M3 <dbl>, CARBUREACTEUR_M3 <dbl>