convert

Try to convert columns according to types sended by the API.

convert(data)

Arguments

data

a dataframe returned by get_data()

Value

a tibble with converted integer and number columns

Details

CAUTION For private life reason, data returns by DiDo are secretize (the value is replaced by the string "secret") so readr can't determine data type.

All secret values will be transform to NA

You can find column's type with columns()

Examples

datafiles() %>%
  dido_search("6c79805c-def9-4695-9d9f-7d86599c4d8a") %>%
  get_data() %>%
  convert()
#> # A tibble: 55 × 10
#>    DEPART…¹ DEPAR…² REGIO…³ REGIO…⁴ ANNEE ESSEN…⁵ GAZOL…⁶ FIOUL…⁷ GPL_M3 CARBU…⁸
#>    <chr>    <chr>   <chr>   <chr>   <chr>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>
#>  1 971      Guadel… 01      Guadel… 2021   118664  190809  246583  20837  107177
#>  2 972      Martin… 02      Martin… 2021   125134  171964  377837  17536   88983
#>  3 973      Guyane  03      Guyane  2021    39528  106097   78464   8826   35147
#>  4 974      La Réu… 04      La Réu… 2021   151834  494126  266309  38040  192811
#>  5 976      Mayotte 06      Mayotte 2021    22575  133408       0  11427    7061
#>  6 971      Guadel… 01      Guadel… 2020   112378  187242  273707  21455   96587
#>  7 972      Martin… 02      Martin… 2020   115415  167795  372927  17777   77281
#>  8 973      Guyane  03      Guyane  2020    26717   82004  106901   9147   28781
#>  9 974      La Réu… 04      La Réu… 2020   129677  445872  202724  38551  168824
#> 10 976      Mayotte 06      Mayotte 2020    19592  124040       0  10852    6044
#> # … with 45 more rows, and abbreviated variable names ¹​DEPARTEMENT_CODE,
#> #   ²​DEPARTEMENT_LIBELLE, ³​REGION_CODE, ⁴​REGION_LIBELLE, ⁵​ESSENCE_M3,
#> #   ⁶​GAZOLE_M3, ⁷​FIOUL_M3, ⁸​CARBUREACTEUR_M3