Get data

Get the data of the last millesime of all the datafiles found in data. All columns are returned as chr (see below).

get_data(data, query = list(), col_types = cols(.default = "c"), concat = TRUE)

Arguments

data

a tibble issued from datafiles() or a dataframe with two columns ridandmillesime`.

query

a query to pass to the API to select columns and filter on values.

col_types

how to convert columns, the default is to use char for all columns cols(.default = "c")

concat

TRUE

Value

If concat is TRUE (default), return a tibble with all data concatenated in one tibble. If concat is FALSE, return a list of tibbles. get_data() returns only chr

Details

You can use convert() to convert number and integer.

For private life reason, data returned by DiDo can be secretize (the value is replaced by the string "secret") so readr can't determine data type.

Examples

# get all columns
datafiles() %>%
  dido_search("drom") %>%
  get_data()
#> # A tibble: 50 × 10
#>    DEPARTEMENT_CODE DEPARTEMENT_LIB… REGION_CODE REGION_LIBELLE ANNEE ESSENCE_M3
#>    <chr>            <chr>            <chr>       <chr>          <chr> <chr>     
#>  1 972              Martinique       02          Martinique     2015  110986    
#>  2 974              La Réunion       04          La Réunion     2019  142206    
#>  3 976              Mayotte          06          Mayotte        2019  20514     
#>  4 971              Guadeloupe       01          Guadeloupe     2017  121076    
#>  5 973              Guyane           03          Guyane         2017  34182     
#>  6 976              Mayotte          06          Mayotte        2016  20025     
#>  7 973              Guyane           03          Guyane         2011  25111     
#>  8 973              Guyane           03          Guyane         2014  32161     
#>  9 974              La Réunion       04          La Réunion     2014  122604    
#> 10 976              Mayotte          06          Mayotte        2011  16176     
#> # … with 40 more rows, and 4 more variables: GAZOLE_M3 <chr>, FIOUL_M3 <chr>,
#> #   GPL_M3 <chr>, CARBUREACTEUR_M3 <chr>
# get only DEPARTEMENT_CODE and ESSENCE_M3 columns
datafiles() %>%
  dido_search("drom") %>%
  get_data(query = c(columns = "DEPARTEMENT_CODE,ESSENCE_M3"))
#> # A tibble: 50 × 2
#>    DEPARTEMENT_CODE ESSENCE_M3
#>    <chr>            <chr>     
#>  1 972              110986    
#>  2 974              142206    
#>  3 976              20514     
#>  4 971              121076    
#>  5 973              34182     
#>  6 976              20025     
#>  7 973              25111     
#>  8 973              32161     
#>  9 974              122604    
#> 10 976              16176     
#> # … with 40 more rows
# get only rows where `DEPARTEMENT_CODE == 971
datafiles() %>%
  dido_search("drom") %>%
  get_data(query = c(DEPARTEMENT_CODE = "eq:971"))
#> # A tibble: 10 × 10
#>    DEPARTEMENT_CODE DEPARTEMENT_LIB… REGION_CODE REGION_LIBELLE ANNEE ESSENCE_M3
#>    <chr>            <chr>            <chr>       <chr>          <chr> <chr>     
#>  1 971              Guadeloupe       01          Guadeloupe     2017  121076    
#>  2 971              Guadeloupe       01          Guadeloupe     2015  127784    
#>  3 971              Guadeloupe       01          Guadeloupe     2020  112378    
#>  4 971              Guadeloupe       01          Guadeloupe     2011  123811    
#>  5 971              Guadeloupe       01          Guadeloupe     2019  128464    
#>  6 971              Guadeloupe       01          Guadeloupe     2012  125695    
#>  7 971              Guadeloupe       01          Guadeloupe     2018  121946    
#>  8 971              Guadeloupe       01          Guadeloupe     2014  112856    
#>  9 971              Guadeloupe       01          Guadeloupe     2016  131545    
#> 10 971              Guadeloupe       01          Guadeloupe     2013  117255    
#> # … with 4 more variables: GAZOLE_M3 <chr>, FIOUL_M3 <chr>, GPL_M3 <chr>,
#> #   CARBUREACTEUR_M3 <chr>